+++ scriptworker/gpg.py
@@ -1,8 +1,12 @@
 #!/usr/bin/env python
 """GPG functions
 """
-
 import gnupg
+import logging
+
+from scriptworker.exceptions import ScriptWorkerTaskException
+
+log = logging.getLogger(__name__)
 
 # map the context.config keys to gnupg.GPG kwarg keys
 GPG_CONFIG_MAPPING = {
@@ -31,4 +35,26 @@ def sign(context, data, **kwargs):
     """Sign `data` with the key `kwargs['keyid']`, or the default key if not specified
     """
     gpg = GPG(context)
-    return gpg.sign(data, **kwargs)
+    return str(gpg.sign(data, **kwargs))
+
+
+def verify_signature(context, signed_data, **kwargs):
+    """Verify `signed_data` with the key `kwargs['keyid']`, or the default key if not specified
+    """
+    log.info("Verifying signature...")
+    gpg = GPG(context)
+    verified = gpg.verify(signed_data, **kwargs)
+    if verified.trust_level is not None and verified.trust_level >= verified.TRUST_FULLY:
+        log.info("Fully trusted signature from {}, {}".format(verified.username, verified.key_id))
+    else:
+        raise ScriptWorkerTaskException("Signature could not be verified!")
+    return verified
+
+
+def get_body(context, signed_data, **kwargs):
+    """Returned the unsigned data from `signed_data`.
+    """
+    gpg = GPG(context)
+    verify_signature(context, signed_data)
+    body = gpg.decrypt(signed_data, **kwargs)
+    return str(body)