+++ scriptworker/worker.py
@@ -8,6 +8,7 @@ import sys
 from scriptworker.poll import find_task, get_azure_urls, update_poll_task_urls
 from scriptworker.config import create_config, read_worker_creds
 from scriptworker.context import Context
+from scriptworker.cot import generate_cot
 from scriptworker.exceptions import ScriptWorkerException
 from scriptworker.log import update_logging_config
 from scriptworker.task import complete_task, reclaim_task, run_task, upload_artifacts, worst_level
@@ -42,15 +43,15 @@ async def run_loop(context, creds_key="credentials"):
                 # write an audit logfile to task_log_dir; copy cot into
                 # artifact_dir/cot ?
                 status = await run_task(context)
-                # TODO generate chain of trust artifact
+                generate_cot(context)
             except ScriptWorkerException as e:
                 status = worst_level(status, e.exit_code)
-                log.error("Hit ScriptWorkerTaskException: {}".format(str(e)))
+                log.error("Hit ScriptWorkerException: {}".format(str(e)))
             try:
                 await upload_artifacts(context)
             except ScriptWorkerException as e:
                 status = worst_level(status, e.exit_code)
-                log.error("Hit ScriptWorkerTaskException: {}".format(str(e)))
+                log.error("Hit ScriptWorkerException: {}".format(str(e)))
             await complete_task(context, status)
             cleanup(context)
             await asyncio.sleep(1)