+++ CHANGELOG.md
@@ -5,8 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 ### Added
 - added `firefox_cot_schema.json` for firefox chain of trust
-- added `scriptworker.gpg` for gpg support
-- added `scriptworker.cot` for chain of trust generation
+- added gpg signature creation + verification
+- added chain of trust generation
 - added `scriptworker.task.worst_level` function for determining overall result of task
 
 ### Changed
@@ -17,6 +17,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 - Functions in `test_config` now ignore existing `TASKCLUSTER_` env vars for a clean testing environment
+- `raise_future_exceptions` no longer throws an exception for an empty list of tasks
 
 ## [0.4.0] - 2016-08-19
 ### Added